requirements.txt
wikipedia-api
python-dotenv
langchain-text-splitters
langchain-community
langgraph
langchain[google-genai]
langchain-google-genai
langchain-pinecone
tqdm
pinecone
sentence-transformers
rank-bm25 
faiss-cpu
streamlit
rouge-score